Move refresh action to nav bar button
- Able to remove refresh button from SeptaDraw as well as it's no longer used in the project
