+++ iSEPTA/iSEPTA/Features/NextToArrive/Maps/NextToArriveDetailViewController.swift
@@ -72,6 +72,23 @@ class NextToArriveDetailViewController: UIViewController, IdentifiableController
         }
     }
 
+    @IBAction func refreshButtonTapped(_ sender: Any) {
+        guard let target = store.state.targetForScheduleActions() else { return }
+        
+        switch target {
+        case .nextToArrive:
+            let action = NextToArriveRefreshDataRequested(refreshUpdateRequested: true)
+            store.dispatch(action)
+        case .favorites:
+            guard var nextToArriveFavorite = store.state.favoritesState.nextToArriveFavorite else { return }
+            nextToArriveFavorite.refreshDataRequested = true
+            let favoriteAction = RequestFavoriteNextToArriveUpdate(favorite: nextToArriveFavorite, description: "User manually refreshed a favorite in Next to Arrive")
+            store.dispatch(favoriteAction)
+        default:
+            break
+        }
+    }
+    
     func favoriteState_NextToArriveFavoriteUpdated(favorite: Favorite?) {
         if let favorite = favorite {
             navigationItem.title = favorite.favoriteName
