+++ iSEPTA/iSEPTA/Features/NextToArrive/nextToArriveDetail.storyboard
@@ -71,7 +71,11 @@
                         <rightBarButtonItems>
                             <barButtonItem systemItem="edit" id="T8u-CP-KOb"/>
                             <barButtonItem image="favoriteNavBarOff" id="4P4-eu-BwU" userLabel="CreateFavorite"/>
-                            <barButtonItem image="refreshIcon" id="Xig-aF-RrS"/>
+                            <barButtonItem image="refreshIcon" id="Xig-aF-RrS">
+                                <connections>
+                                    <action selector="refreshButtonTapped:" destination="tkq-Rq-kOk" id="ccW-Ln-z8U"/>
+                                </connections>
+                            </barButtonItem>
                         </rightBarButtonItems>
                     </navigationItem>
                     <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" barStyle="black" prompted="NO"/>
@@ -286,38 +290,18 @@
                                     <constraint firstAttribute="height" constant="1" id="zgW-rK-imX"/>
                                 </constraints>
                             </view>
-                            <view clearsContextBeforeDrawing="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="42j-gG-sXU" customClass="RefreshNextToArriveDetailButton" customModule="Septa" customModuleProvider="target">
-                                <rect key="frame" x="288" y="20" width="75" height="25"/>
-                                <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
-                                <accessibility key="accessibilityConfiguration" label="Refresh Next To Arrive Data">
-                                    <accessibilityTraits key="traits" button="YES"/>
-                                    <bool key="isElement" value="YES"/>
-                                </accessibility>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="25" id="3r2-5w-ZVj"/>
-                                    <constraint firstAttribute="width" constant="75" id="Qe4-in-bWa"/>
-                                </constraints>
-                                <connections>
-                                    <action selector="refreshButtonTapped:" destination="0La-Ry-f7T" eventType="touchUpInside" id="q1r-ft-5Np"/>
-                                </connections>
-                            </view>
                         </subviews>
                         <color key="backgroundColor" white="0.0" alpha="1" colorSpace="calibratedWhite"/>
                         <constraints>
                             <constraint firstItem="FfQ-gZ-8tW" firstAttribute="leading" secondItem="A9e-oT-VFV" secondAttribute="leading" id="2Mi-JH-ahT" userLabel="DividerLeading"/>
                             <constraint firstItem="8rU-a0-dpY" firstAttribute="baseline" secondItem="8Tw-Og-gvv" secondAttribute="baseline" id="89f-mr-TC7" userLabel="EndNameBaseline"/>
-                            <constraint firstItem="42j-gG-sXU" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="8rU-a0-dpY" secondAttribute="trailing" constant="10" id="8nB-Aa-PfP" userLabel="EndNameRight"/>
                             <constraint firstItem="d62-sN-tBq" firstAttribute="leading" secondItem="f3B-d9-8HH" secondAttribute="trailing" constant="3" id="CVC-SO-apg" userLabel="StartNameLeft"/>
                             <constraint firstItem="FfQ-gZ-8tW" firstAttribute="top" secondItem="A9e-oT-VFV" secondAttribute="top" id="DDG-C9-ehO" userLabel="DividerTop"/>
-                            <constraint firstItem="42j-gG-sXU" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="d62-sN-tBq" secondAttribute="trailing" constant="10" id="DXf-I5-BxC" userLabel="StartLabelRight"/>
-                            <constraint firstItem="42j-gG-sXU" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="8Tw-Og-gvv" secondAttribute="trailing" constant="10" id="JfV-tg-8c1" userLabel="EndLabelRight"/>
                             <constraint firstItem="8rU-a0-dpY" firstAttribute="leading" secondItem="8Tw-Og-gvv" secondAttribute="trailing" constant="3" id="XDS-GF-v80" userLabel="EndNameLeft"/>
                             <constraint firstItem="f3B-d9-8HH" firstAttribute="leading" secondItem="A9e-oT-VFV" secondAttribute="leading" constant="10" id="auc-G6-9C4" userLabel="StartLabelLeft"/>
-                            <constraint firstItem="42j-gG-sXU" firstAttribute="centerY" secondItem="Ixl-7k-2ZF" secondAttribute="centerY" id="eNU-fY-dGa" userLabel="RefreshButtonCenterY"/>
                             <constraint firstItem="d62-sN-tBq" firstAttribute="baseline" secondItem="f3B-d9-8HH" secondAttribute="baseline" id="gD0-Mx-cgd" userLabel="StartNameBaseLine"/>
                             <constraint firstItem="8Tw-Og-gvv" firstAttribute="top" secondItem="f3B-d9-8HH" secondAttribute="bottom" constant="8" id="hqF-5c-suL" userLabel="EndLabelTop"/>
                             <constraint firstItem="8Tw-Og-gvv" firstAttribute="leading" secondItem="A9e-oT-VFV" secondAttribute="leading" constant="10" id="mMN-7K-yGU" userLabel="EndLabelLeft"/>
-                            <constraint firstItem="Ixl-7k-2ZF" firstAttribute="trailing" secondItem="42j-gG-sXU" secondAttribute="trailing" constant="12" id="thY-QA-6hu" userLabel="RefreshButtonRight"/>
                             <constraint firstAttribute="trailing" secondItem="FfQ-gZ-8tW" secondAttribute="trailing" id="wFa-Ok-WLy" userLabel="DividerRight"/>
                             <constraint firstItem="f3B-d9-8HH" firstAttribute="top" secondItem="A9e-oT-VFV" secondAttribute="top" constant="12" id="xVt-EQ-Me5" userLabel="StartLabelTop"/>
                         </constraints>
